Added sentiment try-catch, enum, renamed component to Triton, and sorting by score
