+++ lib/src/main/kotlin/org/promethist/core/model/Sentiment.kt
@@ -1,0 +1,3 @@
+package org.promethist.core.model
+
+enum class Sentiment { NEGATIVE, NEUTRAL, POSITIVE, UNKNOWN }
